Move away tempdir related change from basequeue.
