+++ tests/func/experiments/test_show.py
@@ -524,13 +524,13 @@ def test_show_running_tempdir(tmp_dir, scm, dvc, exp_stage, mocker):
     (tmp_dir / pidfile).dump_json(info.asdict())
     mock_fetch = mocker.patch.object(
         dvc.experiments.tempdir_queue,
-        "_fetch_running_exp",
+        "get_running_exps",
         return_value={exp_rev: info.asdict()},
     )
 
     results = dvc.experiments.show()
     mock_fetch.assert_has_calls(
-        [mocker.call(stash_rev, pidfile, True)],
+        [mocker.call(True)],
     )
     exp_data = get_in(results, [baseline_rev, exp_rev, "data"])
     assert exp_data["status"] == "Running"
